requirements.txt
requests>=2.31.0
google-genai>=1.44.0
anthropic>=0.42.0
python-dotenv>=1.0.1
